pyproject.toml
[tool.poetry]
name = "health-bot-api"
version = "2025.07.0"
description = "Health Bot API"
authors = ["Duc Thanh Nguyen <ngtduc693@gmail.com>"]

packages = [
    { include = "app" }
]

[project]
name = "health-bot-api"
version = "2025.07.0"
description = "Health Bot API"

[tool.poetry.dependencies]
python = ">=3.13"
fastapi = "^0.115.14"
pydantic = "^2.11.7"
python-multipart = "^0.0.20"
email-validator = "^2.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.16.2"
bcrypt = "^4.3.0"
pyjwt = "^2.10.1"
sentry-sdk = {extras = ["fastapi"], version = "^2.32.0"}
azure-cosmos = "^4.9.0"
uv = "^0.7.17"
uvicorn = "^0.35.0"
azure-identity = "^1.23.0"
azure-keyvault-secrets = "^4.10.0"
pydantic-settings = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
coverage = "^7.9.1"
uv = "^0.7.17"
httpx = "^0.28.1"
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core>=2.1.3"]
build-backend = "poetry.core.masonry.api"
